II-S-1: * fix: colors * feature: board generator * feature: validate field * feature: number picker
